reorganize tests in seperate directory
